Deduplicate food/potion select code
